Fix some timeout and server starting/stopping issues
