+++ k8s/service_controller/tests/sb_create_test.go
@@ -3,13 +3,9 @@ package tests
 import (
 	"testing"
 
-	"bytes"
 	"fmt"
-	"io/ioutil"
-	"net/http"
 	"os"
-	"os/exec"
-	"time"
+	"strings"
 )
 
 func TestMain(m *testing.M) {
@@ -18,53 +14,17 @@ func TestMain(m *testing.M) {
 		fmt.Printf("Error starting server: %s\n", err)
 		os.Exit(1)
 	}
-	defer StopServer()
-	os.Exit(m.Run())
+	rc := m.Run()
+	StopServer()
+	os.Exit(rc)
 }
 
-func Test_SB_ping(t *testing.T) {
-	_, err := ServerGET("/v2/service_brokers")
+func Test_SB_Ping(t *testing.T) {
+	res, err := ServerGET("/v2/service_brokers")
 	if err != nil {
-		t.Fatal(err)
+		t.Fatal("Error getting list of brokers: %s", err)
 	}
-}
-
-// Utils stuff
-var serverURL = "http://localhost:10000/"
-
-func ServerGET(url string) (string, error) {
-	resp, err := http.Get(serverURL + url)
-	if err != nil {
-		return "", err
+	if strings.TrimSpace(res) != "[]" {
+		t.Fatal("Should be an empty list: %v", res)
 	}
-	body, err := ioutil.ReadAll(resp.Body)
-	return string(body), err
-}
-
-var cmd *exec.Cmd
-var stdout bytes.Buffer
-var stderr bytes.Buffer
-
-func StartServer() error {
-	cmd = exec.Command("../service_controller")
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	err := cmd.Start()
-	if err != nil {
-		return err
-	}
-
-	// Wait for the server to be available
-	for {
-		if _, err = ServerGET("/v2/service_brokers"); err == nil {
-			break
-		}
-		time.Sleep(1 * time.Second)
-	}
-	return nil
-}
-
-func StopServer() {
-	cmd.Process.Kill()
-	cmd.Wait()
 }
